Added better movement and AI patrolling
Made it so that when the player is coruching, their speed decreases and connot sprint
while coruching, vice versa. Furthermore, Added an AI that patrol between waypoints using
the Navmesh navigation system. Still required optimization as AI needs to pause in an area
to make it more ralistic; there are also glitches where the enemy goes through the walls.