Added a title screen and delay to Enemy AI
By creating another scene, I have created a title screen that allows the player to play the game
and quit the game. The art of the backround was created using chatgpt-4. I have also tweaked the
attack mechanism for the enemy. Previously it would damage the player 0.4 seconds every time it
came to contact with the player. Now as soon as it has contact with the player it will cast an
attack with a delay of 0.4, so if the player is not in distance within the 0.4 second, the attack
will not hit.