added scenes, expanded levels and added doors
I have added several player scenes one of which is a death scene - once a player dies this scene
is shown and allows the player to retry or go to main menu. The other scene is the introduction
which introduces the story to the player. I still require to add the outro scene for when the
player reaches the end of the levels to give them the story ending and then the end scene to
output the users time and allow them to try again, go to main menu or quit. For the expanded
levels I have added an introduction levels, a parkour level and another battle level. Most
levels are seperated by doors whic can be opened by the user.