Fixed animation bug and added more enemies.
There was an animation bug where if two or more enemies were placed, the only
one would have working animations. This bug got changed as it would affect anything that
has the same tag as the enemy, now it is personalised to the attached object. I have
also added more enemies with different pathing.