fixed animation, added doors, signs and levels.
I have fixed an animation error that was still happening with the enemies. I have
also added several sign to teach the controls to the user, alongside with another level
expansion. There is also an end scene with an outro for the story. Need to addd a timer
so that the completion time is outputted at the end.