Added a sensitivity slider, stamina and lighting.
Added player stamina, so that once the player sprints, their stamina get reduced.
If the stamina is depleted, player will stop sprinting and the stamina will
regenerate. Added lighting as well as a ceiling for better visual effect. Lastly,
I added a slider in the pause menu so that the player can adjust it to their needs.